+++ teneto/neuroimagingtools/__init__.py
@@ -1,0 +1,12 @@
+"""Tools to help neuroimaging analyses"""
+
+from .fmriutils import make_parcellation
+from .bidsutils import make_directories, drop_bids_suffix, \
+    get_bids_tag, load_tabular_file, \
+    get_sidecar, confound_matching, \
+    process_exclusion_criteria
+
+__all__ = ['make_parcellation', 'tnet_to_nx', 'make_directories', 'drop_bids_suffix',
+           'get_bids_tag', 'load_tabular_file',
+           'get_sidecar', 'confound_matching',
+           'process_exclusion_criteria']